// src/app/models/pokemon.model.ts
export interface Pokemon {
    id: number;
    sprites: Sprite;
    name: string;
}

interface Sprite {
    back_default?: string;
    back_female?: string;
    back_shiny?: string
    back_shiny_female?: string;
    front_default?: string;
    front_female?: string;
    front_shiny?: string;
    front_shiny_female?: string;
}